requirements.txt
hopsworks[python,great-expectations]==4.2.2
pydantic-settings>=2.6.1
geopy==2.4.1
openmeteo-requests
requests-cache==1.2.0
retry-requests==2.0.0
xgboost==2.0.3
scikit-learn==1.2.2
matplotlib==3.8.3
#plotly
seaborn
nbformat
entsoe-py
pvlib